perf: inline SSSE3 scanners and isolate can_parse
Force-inline the gcc SSSE3 nibble helpers so each 16-byte window
is not a call/ret. Fold the uncommon host checks into one branch,
keep get_href HTTPS-only on the common path, and move can_parse
(now SSSE3 nibble-scanned) out of the unity TU so setter I-cache
is not taxed by the validator.
