+++ .github/workflows/lint_and_format_check.yml
@@ -44,7 +44,7 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y clang-tidy-22 libc++-22-dev libc++abi-22-dev
 
-      # Run clang-tidy on the unity TU and the parser TUs. This keeps
+      # Run clang-tidy on the unity TU, parser TUs, and can_parse. This keeps
       # tests, benchmarks, the generated singleheader amalgamation, and
       # vendored deps out of scope, matching tools/run-clangcldocker.sh.
       # Setting CMAKE_CXX_CLANG_TIDY instead would lint those
@@ -52,6 +52,6 @@ jobs:
       - name: Run clang-tidy
         run: >
           cmake -B build -DADA_USE_UNSAFE_STD_REGEX_PROVIDER=ON -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DCMAKE_CXX_FLAGS="-stdlib=libc++" &&
-          clang-tidy-22 -p build src/ada.cpp src/parser_hot.cpp src/parser_finish.cpp
+          clang-tidy-22 -p build src/ada.cpp src/parser_hot.cpp src/parser_finish.cpp src/can_parse.cpp
         env:
           CXX: clang++-22
